Homework 13

See merge request grodno-java-ee-training-2019/petushok_valiantsin!22

+++ Hotel1/repository-api/src/main/java/com/senlainc/git_courses/java_training/petushok_valiantsin/api/repository/IAttendanceDao.java
@@ -2,6 +2,6 @@ package com.senlainc.git_courses.java_training.petushok_valiantsin.api.repositor
 
 import com.senlainc.git_courses.java_training.petushok_valiantsin.model.Attendance;
 
-public interface IAttendanceDao extends ICommonDao<Attendance, Integer> {
+public interface IAttendanceDao extends ICommonDao<Attendance, Long> {
 
 }
